Renamed PATH to DEPLOY_PATH, made it public, made all the methods in FilesHandeler receive the absolute path, added readFile and doesFileExist, changed variable names and error messages, and made in parseJsonToObject for the fromJson method to get a string and get that string from the readFile method (which will check for .tmp file as well)

+++ src/main/java/frc/trigon/robot/utilities/FilesHandler.java
@@ -3,12 +3,13 @@ package frc.trigon.robot.utilities;
 import edu.wpi.first.wpilibj.Filesystem;
 
 import java.io.File;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 
 public class FilesHandler {
 
-    private static final String PATH = Filesystem.getDeployDirectory().getPath() + "\\";
+    public static final String DEPLOY_PATH = Filesystem.getDeployDirectory().getPath() + "\\";
 
     /**
      * Deletes a file.
@@ -17,9 +18,9 @@ public class FilesHandler {
      * @throws IOException if the method failed to delete the specified file
      */
     public static void deleteFile(String name) throws IOException {
-        File file = new File(PATH + name);
+        File file = new File(name);
         if(!file.delete()) {
-            throw new IOException("Failed to delete file.");
+            throw new IOException("Failed to delete the file \"" + name + "\".");
         }
     }
 
@@ -31,7 +32,7 @@ public class FilesHandler {
      * @throws IOException if the method failed to write the text to the file.
      */
     public static void writeStringToFile(String name, String text) throws IOException {
-        FileWriter fileWriter = new FileWriter(PATH + name);
+        FileWriter fileWriter = new FileWriter(name);
         fileWriter.write(text);
         fileWriter.close();
     }
@@ -39,14 +40,14 @@ public class FilesHandler {
     /**
      * Renames a file.
      *
-     * @param oldName the name of the file to rename
+     * @param name    the name of the file to rename
      * @param newName the new name of the desired name
      * @throws IOException if the method failed to rename the file
      */
-    public static void renameFile(String oldName, String newName) throws IOException {
-        File file = new File(PATH + oldName);
+    public static void renameFile(String name, String newName) throws IOException {
+        File file = new File(name);
         if(!file.renameTo(new File(newName))) {
-            throw new IOException("Failed to rename file " + oldName + " to " + newName);
+            throw new IOException("Failed to rename file " + name + " to " + newName);
         }
     }
 
@@ -62,8 +63,32 @@ public class FilesHandler {
      */
     public static void safeWrite(String text, String name) throws IOException {
         writeStringToFile(name + ".tmp", text);
-        renameFile(name, name + ".bak");
+        if(doesFileExist(name))
+            renameFile(name, name + ".bak");
         renameFile(name + ".tmp", name);
-        deleteFile(name + ".bak");
+        if(doesFileExist(name + ".bak"))
+            deleteFile(name + ".bak");
+    }
+
+    /**
+     * Reads a file and returns its content as a string.
+     * If the files does not exist, it will check for the .tmp file.
+     *
+     * @param name the name of the file to read
+     * @return the file content
+     * @throws IOException if the method failed to read the file
+     */
+    public static String readFile(String name) throws IOException {
+        FileReader reader;
+        StringBuilder fileContent = new StringBuilder();
+        reader = new FileReader(!doesFileExist(name) && doesFileExist(name + ".tmp") ? name + ".tmp" : name);
+        while(reader.read() != -1) {
+            fileContent.append((char) reader.read());
+        }
+        return fileContent.toString();
+    }
+
+    private static boolean doesFileExist(String name) {
+        return new File(name).exists();
     }
 }
